requirements.txt
fastapi==0.95.0
Flask==2.1.2
nebula3_python==3.1.0
pydantic==1.10.5
python-dotenv==1.0.0
PyYAML==6.0
redis==4.3.4
SQLAlchemy==1.4.44
uvicorn==0.21.1
Werkzeug==2.1.2
PyMySQL==1.0.2
puccini==0.21.0